Refactor tabs component and update tab search parameter handling
