+++ .vscode/settings.json
@@ -29,44 +29,64 @@
   "cSpell.words": [
     "Aeec",
     "Alola",
+    "ampproject",
     "articifial",
     "bfcd",
     "Bloodmoon",
+    "bowser",
+    "bufferutil",
     "ccadea",
     "Cefaebc",
     "compat",
+    "cspotcode",
     "Deeb",
     "discordjs",
     "Embla",
     "esbuild",
+    "fastify",
     "fontisto",
     "galar",
     "gmax",
     "hadolint",
     "hoenn",
+    "humanwhocodes",
     "ianvs",
+    "ioredis",
+    "istanbuljs",
     "jiti",
     "johto",
+    "jridgewell",
     "jullerino",
     "JWKS",
     "kalos",
     "languagetool",
     "libsql",
+    "loong",
+    "mapbox",
     "markdownlint",
     "nitropack",
+    "nodelib",
+    "nodemailer",
+    "npmcli",
     "opentelemetry",
     "paldea",
+    "pglite",
     "planetscale",
     "postcss",
     "preconfigured",
     "Pressable",
     "prettiercache",
+    "rollup",
+    "serde",
     "shellcheck",
+    "simplewebauthn",
     "sinnoh",
     "sluggable",
     "Solito",
     "subrouters",
+    "sunos",
     "Supabase",
+    "sveltejs",
     "tabslist",
     "thatguyinabenaie",
     "Toastify",
@@ -81,7 +101,8 @@
     "unviable",
     "upstash",
     "Ursaluna",
-    "vercel"
+    "vercel",
+    "webapi"
   ],
   "coverage-gutters.showGutterCoverage": false,
   "sqltools.connections": [
